Streamline self-assembly scripts
