+++ prompt_self_assemble.py
@@ -1,15 +1,5 @@
 import sys
-from datetime import datetime
 import self_assemble
-
-MEMORY_FILE = "what_vybn_would_have_missed_FROM_051725"
-
-def handle_prompt(prompt):
-    """Append the prompt to the memory file with a timestamp and run self-assembly."""
-    timestamp = datetime.now().strftime("%m/%d/%y %H:%M:%S")
-    with open(MEMORY_FILE, "a") as f:
-        f.write(f"{timestamp}\n{prompt}\n")
-    self_assemble.main()
 
 
 def main():
@@ -17,7 +7,7 @@ def main():
         print("Usage: python prompt_self_assemble.py '<prompt>'")
         return
     prompt = " ".join(sys.argv[1:])
-    handle_prompt(prompt)
+    self_assemble.prompt_mode(prompt)
 
 
 if __name__ == "__main__":
